pyproject.toml
[tool.poetry]
name = "solana-hook"
version = "0.0.1"
description = "Solana Hook"
authors = ["Bevan Hunt <bevan@bevanhunt.com>"]
keywords = ["solana", "blockchain", "web3"]
license = "MIT"

[tool.poetry.dependencies]
python = "==3.12.7"
solana = "==0.35.0"
asyncstdlib = "==3.12.5"
solders = "==0.21.0"
httpx = "==0.27.2"
python-dotenv = "==1.0.1"

[tool.poetry.dev-dependencies]
ruff = "==0.6.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
